Extract apps Vite upload handling
